+++ src/main/java/org/lightjason/agentspeak/action/builtin/math/CMinIndex.java
@@ -74,7 +74,7 @@ public final class CMinIndex extends IBuiltinAction
 
         p_return.add(
             CRawTerm.from(
-                (long) IntStream.range( 0, l_list.size() - 1 )
+                (double) IntStream.range( 0, l_list.size() - 1 )
                                 .parallel()
                                 .reduce( ( i, j ) -> l_list.get( i ) < l_list.get( j ) ? i : j )
                                 .orElseThrow( () -> new CRuntimeException( p_context ) )
