+++ src/main/java/org/lightjason/agentspeak/action/builtin/math/bit/matrix/CColumns.java
@@ -74,7 +74,7 @@ public final class CColumns extends IBuiltinAction
         // arguments are matrix objects
         CCommon.flatten( p_argument )
                .map( ITerm::<BitMatrix>raw )
-               .mapToLong( BitMatrix::columns )
+               .mapToDouble( BitMatrix::columns )
                .boxed()
                .map( CRawTerm::from )
                .forEach( p_return::add );
